+++ src/pyramid_sqlalchemy/fixtures.py
@@ -4,23 +4,33 @@ except ImportError:
     import mock
 import pytest
 from sqlalchemy import create_engine
+from .testing import DatabaseTestCase
 from . import Session
 from . import metadata
 from . import init_sqlalchemy
 
 
+DEFAULT_URI = 'sqlite:///'
+
+
 def pytest_addoption(parser):
-    parser.addoption('--sql-url', default='sqlite:///',
+    parser.addoption('--sql-url', default=DEFAULT_URI,
             help='SQLAlchemy Database URL')
     parser.addoption('--sql-echo', default=False, action='store_true',
             help='Echo SQL statements to console')
 
 
+def pytest_configure(config):
+    DatabaseTestCase.db_uri = config.getoption('sql_url')
+
+
 def pytest_generate_tests(metafunc):
     if 'sqlalchemy_url' in metafunc.fixturenames:
-        metafunc.parametrize('sqlalchemy_url', [metafunc.config.option.sql_url], scope='session')
+        metafunc.parametrize('sqlalchemy_url',
+        [metafunc.config.getoption('sql_url')], scope='session')
     if 'sql_echo' in metafunc.fixturenames:
-        metafunc.parametrize('sql_echo', [metafunc.config.option.sql_echo], scope='session')
+        metafunc.parametrize('sql_echo',
+        [metafunc.config.getoption('sql_echo')], scope='session')
 
 
 @pytest.yield_fixture(scope='session')
@@ -73,4 +83,4 @@ def sql_session(transaction, _sqlalchemy, monkeypatch):
     return _sqlalchemy
 
 
-__all__ = ['transaction', 'sqlalchemy']
+__all__ = ['transaction', 'sql_session']