+++ docs/changes.rst
@@ -4,7 +4,9 @@ Changelog
 1.3 - Unreleased
 --------------------------
 
-- ...
+- Allow overriding the database used for testing with the pytest ``--sql-url`` 
+  option when using the ``DatabaseTestCase`` test class. For non-pytest users
+  support the ``DB_URI`` environment variable as well.
 
 
 1.2.2 - September 11, 2014